event cards in admin dashboard
